backend/requirements.txt

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain and LangGraph dependencies
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.5
langchain-community==0.0.10

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1

# Data processing and utilities
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.0

# HTTP client
httpx==0.25.2
aiofiles==23.2.0

# Logging and monitoring
structlog==23.2.0

# Optional: For enhanced NLP capabilities
nltk==3.8.1
spacy==3.7.2

# Optional: For advanced text similarity
sentence-transformers==2.2.2
